Add support for requiring via npm module loaders
